add removeDuplicates helper function
