+++ src/components/Book/Book.js
@@ -2,9 +2,10 @@ import React, {useState, useEffect} from 'react'
 import { useParams } from 'react-router-dom'
 import { getBookDetails } from '../../api'
 import { removeTags } from '../../helpers'
+import ButtonToRead from '../ButtonToRead/ButtonToRead'
 import './Book.css'
 
-function Book () {
+function Book ({myBooks, addRemove}) {
   const [book, setBook] = useState(null)
   const [isLoading, setIsLoading] = useState(true)
 
@@ -28,6 +29,7 @@ function Book () {
           <img className='book-image' src={book.volumeInfo.imageLinks.thumbnail}></img>
           <p>{book.volumeInfo.title}</p>
           <p>{book.volumeInfo.authors.join(', ')}</p>
+          <ButtonToRead myBooks={myBooks} addRemove={addRemove} book={book}/>
           <p>{removeTags(book.volumeInfo.description)}</p>
         </div>}
   </>
